+++ cmd/sm-config-edit/frontend/wailsjs/go/configedit/App.d.ts
@@ -46,6 +46,8 @@ export function OpenInEditor():Promise<void>;
 
 export function OpenRecent(arg1:string):Promise<configedit.StateDTO>;
 
+export function OpenScriptInEditor(arg1:string):Promise<void>;
+
 export function PreviewAction(arg1:configedit.ItemDTO,arg2:Array<configedit.FieldDTO>,arg3:configedit.ActionDTO):Promise<configedit.ActionPreviewDTO>;
 
 export function PreviewItem(arg1:configedit.ItemDTO,arg2:Array<configedit.FieldDTO>,arg3:Array<configedit.DisplayDTO>,arg4:string):Promise<configedit.PreviewDTO>;
@@ -77,3 +79,5 @@ export function UnlockSecrets(arg1:string,arg2:configedit.SecretsDTO):Promise<vo
 export function ValidateConfig(arg1:configedit.ConfigDTO):Promise<Array<configedit.ValidationIssueDTO>>;
 
 export function ValidateField(arg1:string,arg2:string):Promise<string>;
+
+export function WatchScript(arg1:string):Promise<void>;